requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
gensim>=4.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
wordcloud>=1.9.0
faiss-cpu>=1.7.0
pyLDAvis>=3.4.0
tqdm>=4.66.0
jupyter>=1.0.0
ipywidgets>=8.1.0
sentence-transformers>=2.2.0
